inject a Jedis pool separate from a Sedis pool such that Java users can easily access the Jedis client by using DI.
